# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SpaFlow - Spa Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pricing/calculate:
    post:
      operationId: calculatePrice
      tags: [pricing]
      summary: Calculate price for a rental
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceCalculationInput"
      responses:
        "200":
          description: Price calculation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceCalculation"

  # Check-in
  /checkin:
    post:
      operationId: checkIn
      tags: [clients]
      summary: Full check-in flow - process payment and assign resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInInput"
      responses:
        "200":
          description: Check-in successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckInResult"

  # Waitlist
...
components:
  schemas:
...
    Membership:
      type: object
      required: [id, clientId, type, purchasedAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        type:
          type: string
          enum: [one_time, six_month]
        purchasedAt:
          type: string
          format: date-time
        expiresAt:
          type: ["string", "null"]
          format: date-time
...
    RentalSession:
      type: object
      required: [id, clientId, resourceType, resourceId, resourceName, status, startTime]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        resourceType:
          type: string
          enum: [locker, room]
        resourceId:
          type: integer
        resourceName:
          type: string
        status:
          type: string
          enum: [active, completed, expired]
        startTime:
          type: string
          format: date-time
        expiresAt:
          type: ["string", "null"]
          format: date-time
        endTime:
          type: ["string", "null"]
          format: date-time
        amountPaid:
          type: ["number", "null"]

    # Pricing
    PriceCalculationInput:
      type: object
      required: [clientId, resourceType]
      properties:
        clientId:
          type: integer
        resourceType:
          type: string
          enum: [locker, room]
        membershipType:
          type: ["string", "null"]
          enum: [one_time, six_month, null]
    PriceCalculation:
      type: object
      required: [subtotal, tax, total, appliedRules]
      properties:
        subtotal:
          type: number
        tax:
          type: number
        total:
          type: number
        appliedRules:
          type: array
          items:
            type: string

    # Check-in
    CheckInInput:
      type: object
      required: [clientId, resourceType, resourceId, paymentToken, idempotencyKey]
      properties:
        clientId:
          type: integer
        resourceType:
          type: string
          enum: [locker, room]
        resourceId:
          type: integer
        paymentToken:
          type: string
        idempotencyKey:
          type: string
        membershipType:
          type: ["string", "null"]
          enum: [one_time, six_month, null]
        productIds:
          type: array
          items:
            type: integer
    CheckInResult:
      type: object
      required: [session, transaction]
      properties:
        session:
          $ref: "#/components/schemas/RentalSession"
        transaction:
          $ref: "#/components/schemas/Transaction"
        membership:
          $ref: "#/components/schemas/Membership"

    # Waitlist
...
    Transaction:
      type: object
      required: [id, clientId, amount, type, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        amount:
          type: number
        tax:
          type: number
        total:
          type: number
        type:
          type: string
          enum: [locker_rental, room_rental, membership, product, renewal, extension]
        squarePaymentId:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...